move : move model files to model folder
