--- src/classes/tower.class.js
@@ -1,9 +1,0 @@
-class Tower {
-  constructor(towerId, x, y) {
-    this.id = towerId;
-    this.x = x;
-    this.y = y;
-  }
-}
-
-export default Tower;
